Split GUI shortcut context tests
